Refactors common code into SelectionStrategy Makes SelectionStrategy abstract to incorporate common code.
