+++ benchflow-test-manager/application/src/test/java/cloud/benchflow/testmanager/strategy/selection/OneAtATimeSelectionStrategyTest.java
@@ -28,7 +28,7 @@ public class OneAtATimeSelectionStrategyTest {
   public void setUp() throws Exception {
 
     oneAtATimeSelectionStrategy =
-        new OneAtATimeSelectionStrategy(minioMock, explorationModelDAOMock, testModelDAOMock);
+        new OneAtATimeSelectionStrategy(minioMock, explorationModelDAOMock);
   }
 
   @Test
